macos: start on GL context
Not working or anything at this point, just landing some bits
so that we can do some exploration on how to merge this
Rube Goldberg machine.

What seems difficult right now is to implement the drawing
(such as GdkSurface::render) from the drawInOpenGL
position of the layer. So we'll probably have to come up with
a way to render to an FBO/texture and then draw that from
the layer.