x11/drag: Don't rely on frame events for offscreen windows
The XDND implementation uses an offscreen GdkSurface as IPC when
handling drag'n'drop. This might result in stuck DND, as the IPC window
migth become frozen if the compositing manager doesn't fake those if the
window was never drawn. Avoid this issue by disabling the frame events
handling on the IPC window.