x11: Rename server_time_is_monotonic_time to server_time_uses_monotonic_time
The X server probably uses the monotonic clock when computing its
timestamps, but those timestamps are not just the monotonic clock
time without any sort of changes.  X server timestamps are in
milliseconds and hard truncated to fit into 32-bits.

To make it clear that the server time is derived from, but not
exactly the monotonic time, this commit renames
server_time_is_monotonic_time to server_time_uses_monotonic_time.