Add support for idle inhibition on Wayland
This uses the idle-inhibit protocol from wayland-protocols, to attach an
inhibitor to the GdkSurface.  It can’t do anything if the GtkWindow
isn’t passed, which might warrant some documentation change to encourage
users to not use NULL for this argument.

This implementation prefers the idle-inhibit Wayland protocol to the
DBus version if it is available, since the inhibitor is per-surface
instead of global it allows a finer control over which displays get
dimmed for instance.  For every case not supported by this protocol, a
fallback to the DBus version is used.

This has been tested on Sway.

Fixes #2202.