x11: Don't let compositor process frame unless large enough damage is posted
As of commit 972134abe48a4c9c7b6ad41b0723f30f4e7ae16b, under the
vendor nvidia driver, GTK now inhibits the compositor from
processing frame updates until a fence signals that the GPU is
finished drawing, and until damage is reported from the X server.

It assumes that if the fence has been signaled, that any damage
event after the fence signaling is associated with the fence.
That's, technically, not necessarly always true. The X server could
have generated damage for an unrelated event (say the size of the
window changing), at just the right moment and get in the way.

This commit compensates for that eventuality by copying the painted
region from gdk_x11_gl_context_end_frame, and subtracting the
damaged areas from the copy as they come in.  Once the copied region
goes empty, we know that there won't be any underdraw and can mark
the painting as suitable for the compositor to process.