testsuite: Support different models
In particular, add a GtkTreeModel of GtkDirectoryList that enumerates
the full tree of everything below G_TEST_SRCDIR.

You can set it to ~ or / to get lots of files.

Use this to test numeric sorting, string sorting, the treelistrow
sorter and the multisorter.
And while doing all of that, try to use something that is still
realistic.