Add GskGLShaderNode
This is a rendernode that is supposed to run a GLSL fragment
shader with a set of inputs and produce outputs.
The inputs are:
 * a vec4 args that can be used however the shader wants
 * a list of render nodes that are rendered to textures
 * A glsl string as a GRefString, which will be compiled

Additionally there is a fallback node which is used in case
OpenGL is not supported or there is some kind of failure
with the shader code.