wip: Play with adding a spinner in the search entry
It is not great - FcFontSort is still too choppy to
allow a smooth spinner. And the spinner and the clear
icon interfere with each other a bit: if the icon is
not shown, it makes the spinner appear in the middle
nowhere. This would be better if the spinner was a
builtin feature of the search entry.