gtkfontchooserwidget.c: Fix builds without PangoFT2
Currently, only if PangoFT2 is present and used it is supported to retrieve the
languages that are supported by a particular font, and so we skip building the
parts that populate the language list for filtering and instead filter based on
what the default language is reported to us by Pango.  Don't show the (empty)
language list if we don't have PangoFT2.